New errors on rate fig1
